Fix fibonacci base case across all language implementations (#76)
The recursive implementations in every language returned 1 for both
fibo(0) and fibo(1), so each result was F(x+1) rather than F(x). The
iterative/closed-form variants used the same convention (seeded with
(1, 1) or computed F(n+1)).

Standardise every implementation on F(0)=0, F(1)=1, F(2)=1, ...:

- Recursive base case: return x (or num / *value / etc.) when x < 2,
  instead of returning 1.
- LowLatency.java: short-circuit x<2 and drop the +1 index shift on n.
- Iterative/matrix forms (cpp/loop.cpp, cpp/matrixes.cpp,
  cpp/binpow-matrix.cpp, cpp/interpreter.cpp, go/loop.go,
  haskell/iterate.hs, haskell/object_dynamic.hs,
  haskell/tail_recursion.hs, haskell/zipwith.hs): seed (0, 1) instead
  of (1, 1) and drop +1 shifts.
- cpp/decorators.cpp: First.get() now returns 0 (Second remains 1).

Verified locally that the C++, Rust, Go and Java binaries all print
'7-th Fibonacci number is 13' (was 21) and '0-th Fibonacci number is
0' (was 1), and that 32-th gives 2178309 across all of them, so the
Makefile cross-language consistency check still holds.

The new java/FiboTest unit test now passes.

+++ cpp/objects.cpp
@@ -95,7 +95,7 @@ public:
   ~Fibo() override { delete value; }
   int get() override {
     Int* iff =
-      new If(new Less(new Integer(value), new Integer(2)), new Integer(1),
+      new If(new Less(new Integer(value), new Integer(2)), new Integer(value),
         new Add(new Fibo(new Sub(new Integer(value), new Integer(1))),
           new Fibo(new Sub(new Integer(value), new Integer(2)))));
     const int result = iff->get();
